Set-based demux ref removal check
Reimplemented test of demux ref removal check to read in expected and actual files and put the reference names in dedicated sets. Then checks if the expected name is in the actual set. Also added a requirement that the expected set and actual set have the same length because if they do not, that should point to demux not properly removing references.