Update module_enrich.cpp
Added a check to protect from a seg fault/std::bad_alloc when --low_raw_reads flag is specified. Also added a check which stops the reporting of replicates that did not meet specified raw score constraint or that do not have enriched peptides; this would be because dropping replicates we do not care if they fail the threshold or not.